requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.17.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.20
streamlit-option-menu==0.3.6